12.06

+++ str_tasks/11.js
@@ -6,6 +6,6 @@ let numbers = ['0', '1', '2', '3', '4', '5', '6', '7', '8', '9'];
 let sum = 0;
 
 for (let i = 0; i < str.length; ++i) {
-    if (numbers.includes(str[i])) sum += parseFloat(str[i]);
+    if (numbers.includes(str[i])) sum += parseInt(str[i]);
 }
 console.log(sum);
